Refactor dsl.py
- Extracted relation handling into a new module `relations.py`
- Improved code readability and maintainability in `dsl.py`
- Updated `__init__.py` to reflect new module structure
- Refactored `explorer.py` to use new relation handling

Issue: #34
