+++ buildzr/dsl/__init__.py
@@ -1,3 +1,4 @@
 from .dsl import *
+from .relations import *
 from .explorer import Explorer
 from .expression import Expression
