+++ buildzr/dsl/dsl.py
@@ -29,56 +29,35 @@ from typing import (
 from buildzr.dsl.interfaces import (
     DslWorkspaceElement,
     DslElement,
-    DslRelationship,
-    DslFluentRelationship,
     DslViewsElement,
-    BindLeft,
-    BindLeftLate,
     TSrc, TDst,
     TParent, TChild,
+)
+from buildzr.dsl.relations import (
+    _is_software_container_fluent_relationship,
+    _is_container_component_fluent_relationship,
+    _Relationship,
+    _RelationshipDescription,
+    _FluentRelationship,
+    DslElementRelationOverrides,
 )
 
 def _child_name_transform(name: str) -> str:
     return name.lower().replace(' ', '_')
 
-def _is_software_container_fluent_relationship(
-    obj: '_FluentRelationship[Any, Any]') -> TypeGuard["_FluentRelationship['SoftwareSystem', 'Container']"]:
+# TODO: Remove, not used.
+# def _create_linked_relationship_from(
+#     relationship: '_Relationship[TSrc, TDst]') -> buildzr.models.Relationship:
 
-    return isinstance(obj._parent, SoftwareSystem) and all([
-        isinstance(x, Container) for x in obj._children
-    ])
+#     src = relationship.source
+#     dst = relationship.destination
 
-def _is_container_component_fluent_relationship(
-    obj: '_FluentRelationship[Any, Any]') -> TypeGuard["_FluentRelationship['Container', 'Component']"]:
-
-    return isinstance(obj._parent, Container) and all([
-        isinstance(x, Component) for x in obj._children
-    ])
-
-def _create_linked_relationship_from(
-    relationship: '_Relationship[TSrc, TDst]') -> buildzr.models.Relationship:
-
-    src = relationship.source
-    dst = relationship.destination
-
-    return buildzr.models.Relationship(
-        id=GenerateId.for_relationship(),
-        sourceId=str(src.model.id),
-        destinationId=str(dst.parent.model.id),
-        linkedRelationshipId=relationship.model.id,
-    )
-
-@dataclass
-class With:
-    tags: Optional[Set[str]] = None
-    properties: Optional[Dict[str, str]] = None
-    url: Optional[str] = None
-
-@dataclass
-class _UsesData(Generic[TSrc]):
-    relationship: buildzr.models.Relationship
-    source: TSrc
-
+#     return buildzr.models.Relationship(
+#         id=GenerateId.for_relationship(),
+#         sourceId=str(src.model.id),
+#         destinationId=str(dst.parent.model.id),
+#         linkedRelationshipId=relationship.model.id,
+#     )
 
 TypedModel = TypeVar('TypedModel')
 class TypedDynamicAttribute(Generic[TypedModel]):
@@ -88,257 +67,6 @@ class TypedDynamicAttribute(Generic[TypedModel]):
 
     def __getattr__(self, name: str) -> TypedModel:
         return cast(TypedModel, self._dynamic_attributes.get(name))
-
-def desc(value: str, tech: Optional[str]=None) -> '_RelationshipDescription[Union[Person, SoftwareSystem, Container, Component]]':
-    if tech is None:
-        return _RelationshipDescription(value)
-    else:
-        return _RelationshipDescription(value, tech)
-
-class _RelationshipDescription(Generic[TDst]):
-
-    def __init__(self, description: str, technology: Optional[str]=None) -> None:
-        self._description = description
-        self._technology = technology
-
-    def __rshift__(self, destination: TDst) -> '_UsesFromLate[TDst]':
-        if self._technology:
-            return _UsesFromLate(
-                description=(self._description, self._technology),
-                destination=destination,
-            )
-        return _UsesFromLate(
-            description=self._description,
-            destination=destination,
-        )
-
-class _UsesFromLate(BindLeftLate[TDst]):
-    """
-    This method is used to create a relationship between one source element with
-    multiple destination elements, like so:
-
-    ```python
-    u = Person("user")
-    s1 = SoftwareSystem("software1")
-    s2 = SoftwareSystem("software2")
-
-    # Each element in the following list is a `_UsesFromLate` object.
-    u >> [
-        "Uses" >> s1 | With(tags={"linux", "rules"}),
-        ("Reads from", "SQL") >> s1,
-    ]
-    ```
-
-    This requires late left binding (i.e., the source element is bound after the
-    the destination elements in the list are bounded. This is in contrast to how
-    `_UsesFrom` works, where `u >> "Uses >> s1` binds the source element `u`
-    first (i.e., in `u >> "Uses"` into `_UsesFrom` before finally binding `s1`
-    in `((u >> "Uses) >> s2)`).
-    """
-
-    PossibleSourceType = Union['Person', 'SoftwareSystem', 'Container', 'Component', None]
-
-    @dataclass
-    class _LateBindData:
-        tags: Optional[Set[str]] = None
-        properties: Optional[Dict[str, str]] = None
-        url: Optional[str] = None
-
-    def __init__(self, description: Union[str, Tuple[str, str]], destination: TDst) -> None:
-        if isinstance(description, str):
-            self._description = description
-            self._technology: Optional[str] = None
-        elif isinstance(description, tuple) and len(description) == 2:
-            self._description = description[0]
-            self._technology = description[1]
-        self._source: Optional[_UsesFromLate.PossibleSourceType] = None
-        self._destination = destination
-        self._relationship: Optional[_Relationship[_UsesFromLate.PossibleSourceType, TDst]] = None
-        self._late_bind_data: _UsesFromLate._LateBindData = _UsesFromLate._LateBindData()
-
-    def set_source(self, source: PossibleSourceType) -> None:
-        self._source = source
-        self._relationship =  _Relationship(
-            uses_data=_UsesData(
-                relationship=buildzr.models.Relationship(
-                    id=GenerateId.for_relationship(),
-                    description=self._description,
-                    technology=self._technology,
-                    sourceId=str(self._source.model.id),
-                ),
-                source=self._source,
-            ),
-            destination=self._destination,
-        )
-        self._late_bind_with()
-
-    def get_relationship(self) -> Optional['_Relationship[PossibleSourceType, TDst]']:
-        return self._relationship
-
-    def _late_bind_with(self) -> None:
-        """
-        Binds tags, properties, url to the relationship.
-        Called once the relationship is set.
-        """
-        self._relationship = self._relationship.has(
-            tags=self._late_bind_data.tags,
-            properties=self._late_bind_data.properties,
-            url=self._late_bind_data.url,
-        )
-
-
-    def __or__(self, other: With) -> Self:
-        self._late_bind_data.tags = other.tags
-        self._late_bind_data.properties = other.properties
-        self._late_bind_data.url = other.url
-
-        return self
-
-class _UsesFrom(BindLeft[TSrc, TDst]):
-
-    def __init__(self, source: TSrc, description: str="", technology: str="") -> None:
-        self.uses_data = _UsesData(
-            relationship=buildzr.models.Relationship(
-                id=GenerateId.for_relationship(),
-                description=description,
-                technology=technology,
-                sourceId=str(source.model.id),
-            ),
-            source=source,
-        )
-
-    def __rshift__(self, destination: TDst) -> '_Relationship[TSrc, TDst]':
-        if isinstance(destination, Workspace):
-            raise TypeError(f"Unsupported operand type for >>: '{type(self).__name__}' and {type(destination).__name__}")
-        return _Relationship(self.uses_data, destination)
-
-class _Relationship(DslRelationship[TSrc, TDst]):
-
-    @property
-    def model(self) -> buildzr.models.Relationship:
-        return self._m
-
-    @property
-    def tags(self) -> Set[str]:
-        return self._tags
-
-    @property
-    def source(self) -> DslElement:
-        return self._src
-
-    @property
-    def destination(self) -> DslElement:
-        return self._dst
-
-    def __init__(
-            self,
-            uses_data: _UsesData[TSrc],
-            destination: TDst,
-            tags: Set[str]=set(),
-            _include_in_model: bool=True,
-        ) -> None:
-
-        self._m = uses_data.relationship
-        self._tags = {'Relationship'}.union(tags)
-        self._src = uses_data.source
-        self._dst = destination
-        self.model.tags = ','.join(self._tags)
-
-        uses_data.relationship.destinationId = str(destination.model.id)
-
-        if not isinstance(uses_data.source.model, buildzr.models.Workspace):
-            uses_data.source.destinations.append(self._dst)
-            self._dst.sources.append(self._src)
-            if _include_in_model:
-                if uses_data.source.model.relationships:
-                    uses_data.source.model.relationships.append(uses_data.relationship)
-                else:
-                    uses_data.source.model.relationships = [uses_data.relationship]
-
-        # Used to pass the `_UsesData` object as reference to the `__or__`
-        # operator overloading method.
-        self._ref: Tuple[_UsesData] = (uses_data,)
-
-    def __or__(self, _with: With) -> Self:
-        return self.has(
-            tags=_with.tags,
-            properties=_with.properties,
-            url=_with.url,
-        )
-
-    def has(
-        self,
-        tags: Optional[Set[str]]=None,
-        properties: Optional[Dict[str, str]]=None,
-        url: Optional[str]=None,
-    ) -> Self:
-        """
-        Adds extra info to the relationship.
-
-        This can also be achieved using the syntax sugar `DslRelationship |
-        With(...)`.
-        """
-        if tags:
-            self._tags = self._tags.union(tags)
-            self._ref[0].relationship.tags = ",".join(self._tags)
-        if properties:
-            self._ref[0].relationship.properties = properties
-        if url:
-            self._ref[0].relationship.url = url
-        return self
-
-class _FluentRelationship(DslFluentRelationship[TParent, TChild]):
-
-    """
-    A hidden class used in the fluent DSL syntax after specifying a model (i.e.,
-    Person, Software System, Container) to define relationship(s) within the
-    specified model.
-    """
-
-    def __init__(self, parent: TParent, children: Tuple[TChild, ...]) -> None:
-        self._children: Tuple[TChild, ...] = children
-        self._parent: TParent = parent
-
-    def where(
-        self,
-        func: Callable[
-            [TParent],
-            Sequence[
-                Union[
-                    DslRelationship,
-                    Sequence[DslRelationship]
-                ]
-            ]
-        ], implied: bool=False) -> TParent:
-
-        relationships: Sequence[DslRelationship] = []
-
-        func = cast(Callable[[TParent], Sequence[Union[DslRelationship, Sequence[DslRelationship]]]], func)
-
-        # Flatten the resulting relationship list.
-        relationships = [
-            rel for sublist in func(self._parent)
-            for rel in (
-                sublist if isinstance(sublist, list) else [sublist]
-            )
-        ]
-
-        # If we have relationship s >> do >> a.b, then create s >> do >> a.
-        # If we have relationship s.ss >> do >> a.b.c, then create s.ss >> do >> a.b and s.ss >> do >> a.
-        # And so on...
-        if implied:
-            for relationship in relationships:
-                source = relationship.source
-                parent = relationship.destination.parent
-                while parent is not None and not isinstance(parent, DslWorkspaceElement):
-                    r = source.uses(parent, description=relationship.model.description, technology=relationship.model.technology)
-                    r.model.linkedRelationshipId = relationship.model.id
-                    parent = parent.parent
-
-        return self._parent
-
-    def get(self) -> TParent:
-        return self._parent
 
 class Workspace(DslWorkspaceElement):
     """
@@ -448,12 +176,10 @@ class Workspace(DslWorkspaceElement):
     def __dir__(self) -> Iterable[str]:
         return list(super().__dir__()) + list(self._dynamic_attrs.keys())
 
-class SoftwareSystem(DslElement):
+class SoftwareSystem(DslElementRelationOverrides):
     """
     A software system.
     """
-
-    _Affectee = Union['Person', 'SoftwareSystem', 'Container', 'Component']
 
     @property
     def model(self) -> buildzr.models.SoftwareSystem:
@@ -478,27 +204,6 @@ class SoftwareSystem(DslElement):
     @property
     def tags(self) -> Set[str]:
         return self._tags
-
-    def uses(
-        self,
-        other: DslElement,
-        description: Optional[str]=None,
-        technology: Optional[str]=None,
-        tags: Set[str]=set()) -> _Relationship[Self, DslElement]:
-
-        source = self
-
-        uses_from = _UsesFrom[Self, DslElement](
-            source=source,
-            description=description,
-            technology=technology
-        )
-
-        return _Relationship(
-            uses_from.uses_data,
-            destination=other,
-            tags=tags,
-        )
 
     def __init__(self, name: str, description: str="", tags: Set[str]=set(), properties: Dict[str, Any]=dict()) -> None:
         self._m = buildzr.models.SoftwareSystem()
@@ -541,45 +246,6 @@ class SoftwareSystem(DslElement):
     def container(self) -> TypedDynamicAttribute['Container']:
         return TypedDynamicAttribute['Container'](self._dynamic_attrs)
 
-    @overload # type: ignore[override]
-    def __rshift__(self, description_and_technology: Tuple[str, str]) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, description: str) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, _RelationshipDescription: _RelationshipDescription[_Affectee]) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, multiple_destinations: List[_UsesFromLate[_Affectee]]) -> List[_Relationship[Self, _Affectee]]:
-        ...
-
-    def __rshift__(
-            self,
-            other: Union[
-                str,
-                Tuple[str, str],
-                _RelationshipDescription[_Affectee],
-                List[_UsesFromLate[_Affectee]]
-            ]) -> Union[_UsesFrom[Self, _Affectee], List[_Relationship[Self, _Affectee]]]:
-        if isinstance(other, str):
-            return _UsesFrom(self, other)
-        elif isinstance(other, tuple):
-            return _UsesFrom(self, description=other[0], technology=other[1])
-        elif isinstance(other, _RelationshipDescription):
-            return _UsesFrom(self, description=other._description, technology=other._technology)
-        elif isinstance(other, list):
-            relationships = []
-            for dest in other:
-                dest.set_source(self)
-                relationships.append(dest.get_relationship())
-            return cast(List[_Relationship[Self, SoftwareSystem._Affectee]], relationships)
-        else:
-            raise TypeError(f"Unsupported operand type for >>: '{type(self).__name__}' and {type(other).__name__}")
-
     def __getattr__(self, name: str) -> 'Container':
         return self._dynamic_attrs[name]
 
@@ -589,12 +255,10 @@ class SoftwareSystem(DslElement):
     def __dir__(self) -> Iterable[str]:
         return list(super().__dir__()) + list(self._dynamic_attrs.keys())
 
-class Person(DslElement):
+class Person(DslElementRelationOverrides):
     """
     A person who uses a software system.
     """
-
-    _Affectee = Union['Person', 'SoftwareSystem', 'Container', 'Component']
 
     @property
     def model(self) -> buildzr.models.Person:
@@ -624,27 +288,6 @@ class Person(DslElement):
     def tags(self) -> Set[str]:
         return self._tags
 
-    def uses(
-        self,
-        other: DslElement,
-        description: Optional[str]=None,
-        technology: Optional[str]=None,
-        tags: Set[str]=set()) -> _Relationship[Self, DslElement]:
-
-        source = self
-
-        uses_from = _UsesFrom[Self, DslElement](
-            source=source,
-            description=description,
-            technology=technology
-        )
-
-        return _Relationship(
-            uses_from.uses_data,
-            destination=other,
-            tags=tags,
-        )
-
     def __init__(self, name: str, description: str="", tags: Set[str]=set(), properties: Dict[str, Any]=dict()) -> None:
         self._m = buildzr.models.Person()
         self._parent: Optional[Workspace] = None
@@ -658,52 +301,10 @@ class Person(DslElement):
         self.model.tags = ','.join(self._tags)
         self.model.properties = properties
 
-    @overload # type: ignore[override]
-    def __rshift__(self, description_and_technology: Tuple[str, str]) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, description: str) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, _RelationshipDescription: _RelationshipDescription[_Affectee]) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, multiple_destinations: List[_UsesFromLate[_Affectee]]) -> List[_Relationship[Self, _Affectee]]:
-        ...
-
-    def __rshift__(
-        self,
-        other: Union[
-            str,
-            Tuple[str, str],
-            _RelationshipDescription[_Affectee],
-            List[_UsesFromLate[_Affectee]]]
-        ) -> Union[_UsesFrom[Self, _Affectee], List[_Relationship[Self, _Affectee]]]:
-
-        if isinstance(other, str):
-            return _UsesFrom(self, other)
-        elif isinstance(other, tuple):
-            return _UsesFrom(self, description=other[0], technology=other[1])
-        elif isinstance(other, _RelationshipDescription):
-            return _UsesFrom(self, description=other._description, technology=other._technology)
-        elif isinstance(other, list):
-            relationships = []
-            for dest in other:
-                dest.set_source(self)
-                relationships.append(dest.get_relationship())
-            return cast(List[_Relationship[Self, Person._Affectee]], relationships)
-        else:
-            raise TypeError(f"Unsupported operand type for >>: '{type(self).__name__}' and {type(other).__name__}")
-
-class Container(DslElement):
+class Container(DslElementRelationOverrides):
     """
     A container (something that can execute code or host data).
     """
-
-    _Affectee = Union[Person, SoftwareSystem, 'Container', 'Component']
 
     @property
     def model(self) -> buildzr.models.Container:
@@ -728,27 +329,6 @@ class Container(DslElement):
     @property
     def tags(self) -> Set[str]:
         return self._tags
-
-    def uses(
-        self,
-        other: DslElement,
-        description: Optional[str]=None,
-        technology: Optional[str]=None,
-        tags: Set[str]=set()) -> _Relationship[Self, DslElement]:
-
-        source = self
-
-        uses_from = _UsesFrom[Self, DslElement](
-            source=source,
-            description=description,
-            technology=technology
-        )
-
-        return _Relationship(
-            uses_from.uses_data,
-            destination=other,
-            tags=tags,
-        )
 
     def contains(self, *components: 'Component') -> _FluentRelationship['Container', 'Component']:
         if not self.model.components:
@@ -779,45 +359,6 @@ class Container(DslElement):
     def component(self) -> TypedDynamicAttribute['Component']:
         return TypedDynamicAttribute['Component'](self._dynamic_attrs)
 
-    @overload # type: ignore[override]
-    def __rshift__(self, description_and_technology: Tuple[str, str]) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, description: str) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, _RelationshipDescription: _RelationshipDescription[_Affectee]) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, multiple_destinations: List[_UsesFromLate[_Affectee]]) -> List[_Relationship[Self, _Affectee]]:
-        ...
-
-    def __rshift__(
-            self,
-            other: Union[
-                str,
-                Tuple[str, str],
-                _RelationshipDescription[_Affectee],
-                List[_UsesFromLate[_Affectee]]
-            ]) -> Union[_UsesFrom[Self, _Affectee], List[_Relationship[Self, _Affectee]]]:
-        if isinstance(other, str):
-            return _UsesFrom(self, other)
-        elif isinstance(other, tuple):
-            return _UsesFrom(self, description=other[0], technology=other[1])
-        elif isinstance(other, _RelationshipDescription):
-            return _UsesFrom(self, description=other._description, technology=other._technology)
-        elif isinstance(other, list):
-            relationships = []
-            for dest in other:
-                dest.set_source(self)
-                relationships.append(dest.get_relationship())
-            return cast(List[_Relationship[Self, Container._Affectee]], relationships)
-        else:
-            raise TypeError(f"Unsupported operand type for >>: '{type(self).__name__}' and {type(other).__name__}")
-
     def __getattr__(self, name: str) -> 'Component':
         return self._dynamic_attrs[name]
 
@@ -827,12 +368,10 @@ class Container(DslElement):
     def __dir__(self) -> Iterable[str]:
         return list(super().__dir__()) + list(self._dynamic_attrs.keys())
 
-class Component(DslElement):
+class Component(DslElementRelationOverrides):
     """
     A component (a grouping of related functionality behind an interface that runs inside a container).
     """
-
-    _Affectee = Union[Person, SoftwareSystem, Container, 'Component']
 
     @property
     def model(self) -> buildzr.models.Component:
@@ -858,27 +397,6 @@ class Component(DslElement):
     def tags(self) -> Set[str]:
         return self._tags
 
-    def uses(
-        self,
-        other: DslElement,
-        description: Optional[str]=None,
-        technology: Optional[str]=None,
-        tags: Set[str]=set()) -> _Relationship[Self, DslElement]:
-
-        source = self
-
-        uses_from = _UsesFrom[Self, DslElement](
-            source=source,
-            description=description,
-            technology=technology
-        )
-
-        return _Relationship(
-            uses_from.uses_data,
-            destination=other,
-            tags=tags,
-        )
-
     def __init__(self, name: str, description: str="", technology: str="", tags: Set[str]=set(), properties: Dict[str, Any]=dict()) -> None:
         self._m = buildzr.models.Component()
         self._parent: Optional[Container] = None
@@ -892,45 +410,6 @@ class Component(DslElement):
         self.model.relationships = []
         self.model.tags = ','.join(self._tags)
         self.model.properties = properties
-
-    @overload # type: ignore[override]
-    def __rshift__(self, description_and_technology: Tuple[str, str]) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, description: str) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, _RelationshipDescription: _RelationshipDescription[_Affectee]) -> _UsesFrom[Self, _Affectee]:
-        ...
-
-    @overload
-    def __rshift__(self, multiple_destinations: List[_UsesFromLate[_Affectee]]) -> List[_Relationship[Self, _Affectee]]:
-        ...
-
-    def __rshift__(
-            self,
-            other: Union[
-                str,
-                Tuple[str, str],
-                _RelationshipDescription[_Affectee],
-                List[_UsesFromLate[_Affectee]]
-            ]) -> Union[_UsesFrom[Self, _Affectee], List[_Relationship[Self, _Affectee]]]:
-        if isinstance(other, str):
-            return _UsesFrom(self, other)
-        elif isinstance(other, tuple):
-            return _UsesFrom(self, description=other[0], technology=other[1])
-        elif isinstance(other, _RelationshipDescription):
-            return _UsesFrom(self, description=other._description, technology=other._technology)
-        elif isinstance(other, list):
-            relationships = []
-            for dest in other:
-                dest.set_source(self)
-                relationships.append(dest.get_relationship())
-            return cast(List[_Relationship[Self, Component._Affectee]], relationships)
-        else:
-            raise TypeError(f"Unsupported operand type for >>: '{type(self).__name__}' and {type(other).__name__}")
 
 _RankDirection = Literal['tb', 'bt', 'lr', 'rl']
